DROTH-3466 moved road number enumerations to companion object, added rest of number ranges. Deleted RoadAddressUtils.scala
