+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/util/GeometryTransform.scala
@@ -70,6 +70,29 @@ object RoadSide {
 }
 
 case class RoadAddress(municipalityCode: Option[String], road: Int, roadPart: Int, track: Track, addrM: Int)
+object RoadAddress {
+  lazy val routeRoadNumbers: Seq[Int] = 1 to 39
+  lazy val mainRoadRoadNumbers: Seq[Int] =  40 to 99
+  lazy val connectingRoadRoadNumbers: Seq[Int] = 1000 to 19999
+  lazy val roundAboutAndDropletRoadNumbers: Seq[Int] = 20001 to 39999 //road part 1
+  lazy val roundAboutFreeRightSideRoadNumbers: Seq[Int] = 20001 to 39999 //road part 2-5
+  lazy val roundAboutDriveThroughLineRoadNumbers: Seq[Int] = 20001 to 39999 //road part 6-7
+  lazy val rampsAndJunctionsRoadNumbers: Seq[Int] = 20001 to 39999 // road part > 11
+  lazy val serviceAreaRoadNumbers: Seq[Int] = 20001 to 39999 // road part 995-999
+  lazy val serviceAccessRoadNumbers: Seq[Int] = 30000 to 39999 // road part 9
+  lazy val streetRoadNumbers: Seq[Int] = 40001 to 49999
+  lazy val privateRoadRoadNumbers: Seq[Int] = 50001 to 59999
+  lazy val winterRoadRoadNumbers: Seq[Int] = 60001 to 61999
+  lazy val tractorRoadRoadNumbers: Seq[Int] = 62001 to 62999
+  lazy val walkingCyclingRoadNumbers: Seq[Int] = 70001 to 99999
+
+  lazy val carTrafficRoadNumbers: Seq[Int] = 1 to 62999
+
+  def isCarTrafficRoadAddress(roadNumber: Long): Boolean = {
+    carTrafficRoadNumbers.contains(roadNumber)
+  }
+
+}
 class RoadAddressException(response: String) extends RuntimeException(response)
 class RoadPartReservedException(response: String) extends RoadAddressException(response)
